reorganize postponed annotations tests
